requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asttokens==3.0.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
blis==1.2.0
Brlapi==0.8.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.6
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
confection==0.1.5
coverage==7.7.1
cryptography==3.4.8
cupshelpers==1.0
cymem==2.0.11
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
duplicity==0.8.21
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.14.1
Flask==3.1.0
frozenlist==1.2.0
future==0.18.2
gemini_ai==0.0.5
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.1.0
ipython==8.34.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.4.2
jupyterlab_widgets==3.0.13
keyring==23.5.0
langcodes==3.5.0
language-selector==0.1
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==5.1.0
murmurhash==1.0.12
netifaces==0.11.0
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==2.2.3
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.3
proton-vpn-gtk-app==4.9.5
proton-vpn-network-manager==0.12.13
protonvpn-cli==2.2.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.3.5
pytest-cov==6.0.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-telegram-bot==22.0
pythondialog==3.5.3
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2024.11.6
reportlab==3.6.8
requests==2.25.1
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
sentry-sdk==1.4.3
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
systemd-python==234
thinc==8.3.4
threadpoolctl==3.6.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.13.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
